+++ README.md
@@ -10,7 +10,7 @@
 - **masks** - модуль содержит функции для маскировки номеров карт или счетов
 - **processing** модуль обработки банковских операций
 - **widget** модуль содержит различные функции виджета
-
+- **decorators** модуль содержит декораторы для функций
 
 ### 2. tests
 Это пакет с тестами.
