+++ src/config.py
@@ -1,3 +1,3 @@
-from pathlib  import Path
+from pathlib import Path
 
 ROOT_PATH = Path(__file__).parent
